+++ CHANGES.md
@@ -17,6 +17,7 @@
 - PS3: Enable Cheevos.
 - PSP: Enable threading support through pthreads.
 - WINDOWS: Improved Unicode support (for cores/directory creation).
+- WINDOWS: Show progress meter on taskbar for downloads (Windows 7 and up).
 - WIIU: Overlay support.
 - WIIU: Transparency support in menu + overlays.
 - WIIU: New HID pad support.